企业用户删除技术成果专利时移除存放图片的文件夹功能完成

+++ sql/deleteTech.php
@@ -1,30 +1,60 @@
 <?php
+session_start();
+define('FILE_UPLOAD_PATH', "../user_files/");
 include("connection.php");
-
 $url = "../mytech.php";
 $id = $_GET["id"];
+$user_id = $_SESSION['userid'];
 
 /* 删除数据sql */
-if (!empty($id)) {
+if (!empty($id) && !empty($user_id)) {
     try {
         $sql = "delete from `solar_technologies` where `id` like $id";
 
         $pdo->beginTransaction();
         $result = $pdo->prepare($sql);
+
+        //先执行语句删除数据库中的条目
         if ($result->execute()) {
-            echo "<script> alert('删除成功！！');</script>";
-            echo "<meta http-equiv=\"refresh\" content=\"0.5;url=$url\">";
+
+            $dir = FILE_UPLOAD_PATH . $user_id ."/solartech/" . $id;
+
+            //然后删除用户对应的
+            if(is_dir($dir)) {
+                $files = array_diff(scandir($dir), array('.','..'));
+
+                foreach ($files as $file) {
+                    unlink("$dir/$file");
+                }
+
+                //删除成功则提交事务
+                if (rmdir($dir)) {
+                    $pdo->commit();
+                    echo "<script> alert('删除成功！！');</script>";
+                    echo "<meta http-equiv=\"refresh\" content=\"0.5;url=$url\">";
+                }
+
+                //删除失败则回滚
+                else {
+                    $pdo->rollBack();
+                    echo "<script> alert('删除失败！！');</script>";
+//                    echo "<meta http-equiv=\"refresh\" content=\"0.5;url=$url\">";
+                }
+            }
+
         } else {
             echo "<script> alert('删除失败！！');</script>";
-            echo "<meta http-equiv=\"refresh\" content=\"0.5;url=$url\">";
+            echo $pdo->errorInfo();
+//            echo "<meta http-equiv=\"refresh\" content=\"0.5;url=$url\">";
         }
-        $pdo->commit();
+
     } catch (PDOException $e) {
         die("Error!!: " . $e->getMessage() . "<br>");
-        $pdo->rollBack();
     }
 } else {
-    echo "没有接收到ID！！！";
+    echo "没有接收到ID或用户ID为空！！！" . "<br>";
+    echo "user id : " . $user_id . "<br>";
+    echo "tech id : " . $id . "<br>";
 }
 
 ?>
